+++ NEWS.md
@@ -1,3 +1,9 @@
+triplot 0.2
+----------------------------------------------------------------
+* aspectImportance is renamed to triplot 
+* added second vignette
+
+
 aspectImportance 0.1
 ----------------------------------------------------------------
 * aspect_importance is moved out of `DALEXtra` into standalone package